control-service/requirements.txt
grpcio==1.65.5
grpcio-tools==1.65.5
protobuf==5.27.3
python-dotenv==1.0.1
requests==2.32.3
prophet==1.1.6
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

